+++ kafka-functional-tests/src/test/java/com/playtika/shepherd/KafkaPushFarmTest.java
@@ -11,7 +11,7 @@ import java.time.Duration;
 import java.util.List;
 import java.util.Objects;
 import java.util.Random;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Stream;
 
@@ -111,7 +111,7 @@ public class KafkaPushFarmTest extends BasicKafkaTest{
         int ver1 = nextVersion(0, versioned);
 
         AtomicReference<List<ByteBuffer>> cows1 = new AtomicReference<>(List.of());
-        AtomicInteger version1 = new AtomicInteger();
+        AtomicLong version1 = new AtomicLong();
         String herdName = versioned ? "push-dynamic-group-versioned" : "push-dynamic-group";
         Pasture<ByteBuffer> pasture1 = kafkaRanch.addPasture(herdName, (population, version, generation, isLeader) -> {
             logPopulation(1, population, version, isLeader);
@@ -122,7 +122,7 @@ public class KafkaPushFarmTest extends BasicKafkaTest{
         pasture1.start();
 
         AtomicReference<List<ByteBuffer>> cows2 = new AtomicReference<>(List.of());
-        AtomicInteger version2 = new AtomicInteger();
+        AtomicLong version2 = new AtomicLong();
         Pasture<ByteBuffer> pasture2 = kafkaRanch.addPasture(herdName, (population, version, generation, isLeader) -> {
             logPopulation(2, population, version, isLeader);
             cows2.set(population);
@@ -326,7 +326,7 @@ public class KafkaPushFarmTest extends BasicKafkaTest{
         });
     }
 
-    private static void logPopulation(int pastureIndex, List<ByteBuffer> population, int version, boolean isLeader) {
+    private static void logPopulation(int pastureIndex, List<ByteBuffer> population, long version, boolean isLeader) {
         logger.info("Assigned to pasture{} leader={} version={} [{}]", pastureIndex, isLeader, version, toBytes(population));
     }
